get_x_val rework

+++ VM/includes/corewar.h
@@ -151,7 +151,7 @@ void				ft_init_war(t_parse_file file, t_war *war);
 void				init_tab(t_opp *opp_tab);
 int					ft_atoi_base(const char *str, int base);
 char				*ft_itoa_base(int value, int base);
-int		            get_2_val(t_war *war, t_chariot *chariot, int i);
+short		            get_2_val(t_war *war, t_chariot *chariot, int i);
 int		            get_4_val(t_war *war, t_chariot *chariot, int i);
 void        		print_verbose_16(t_war *war, t_chariot *chariot, int size);
 int					read_arena(t_war *war, int cell);
