+++ VM/src/operations/st_fct.c
@@ -27,7 +27,7 @@ int			st_fct(t_war *war, t_chariot *chariot)
 	{
 		param = get_2_val(war, chariot, i);
 		if (war->verbose[2] == 1)
-        	printf("P %4d | st r%d %d\n", (chariot->index + 1), (unsigned char)war->arena[pos + 2], (short)param);
+        	printf("P %4d | st r%d %hd\n", (chariot->index + 1), (unsigned char)war->arena[pos + 2], param);
         print_verbose_16(war, chariot, 5);
 		write_on_arena(war, chariot->registres[(unsigned char)war->arena[pos + 2] - 1], (pos + (param % IDX_MOD)), 4);
 	}
